updating tooltip styles
